requirements.txt
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2022.6.15
charset-normalizer==2.1.0
flake8==4.0.1
frozenlist==1.3.0
idna==3.3
mccabe==0.6.1
multidict==6.0.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.1
scheduler==0.7.4
typeguard==2.13.3
yarl==1.7.2
